requirements.txt
langgraph
langsmith
langchain-openai
pandas
python-dotenv
langsmith 
openai 
langchain_core
cognee